requirements.txt
streamlit
langchain-community
langchain-openai
duckduckgo-search
openai
chromadb
aiofiles
trafilatura
openai
pymupdf
openpyxl
python-docx
pandas